dash-w-logging/requirements.txt
ansi2html==1.8.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.2
Flask==2.2.5
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
MarkupSafe==2.1.3
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.1
plotly==5.18.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
rsconnect-python==1.20.0
semver==2.13.0
six==1.16.0
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.17.0
zope.interface==6.1
